fix: extract user input to cli.rs file
